+++ internal/generate/olm-catalog/package_manifest.go
@@ -28,10 +28,14 @@ import (
 	"github.com/operator-framework/operator-sdk/internal/generate/gen"
 	"github.com/operator-framework/operator-sdk/internal/scaffold"
 	"github.com/operator-framework/operator-sdk/internal/util/fileutil"
+	"github.com/operator-framework/operator-sdk/internal/util/projutil"
 
 	"github.com/ghodss/yaml"
 	log "github.com/sirupsen/logrus"
 )
+
+// Deprecated: The package manifest generator will no longer create new package
+// manifests, only update existing ones. This generator will be removed in v0.19.0.
 
 const (
 	packageManifestFileExt = ".package.yaml"
@@ -94,7 +98,7 @@ func (g pkgGenerator) Generate() error {
 		return err
 	}
 	if len(fileMap) == 0 {
-		return errors.New("error generating package manifest: no generated file found")
+		return nil
 	}
 	pkgManifestOutputDir := filepath.Join(g.OutputDir, OLMCatalogChildDir, g.OperatorName)
 	if err = os.MkdirAll(pkgManifestOutputDir, fileutil.DefaultDirFileMode); err != nil {
@@ -113,17 +117,30 @@ func (g pkgGenerator) Generate() error {
 // generate either reads an existing package manifest or creates a new
 // manifest and modifies it based on values set in s.
 func (g pkgGenerator) generate() (map[string][]byte, error) {
-	pkg, err := g.buildPackageManifest()
-	if err != nil {
-		return nil, err
+	pkgManifestOutputDir := filepath.Join(g.OutputDir, OLMCatalogChildDir, g.OperatorName)
+	path := filepath.Join(pkgManifestOutputDir, g.fileName)
+	pkg := registry.PackageManifest{}
+	if _, err := os.Stat(path); os.IsNotExist(err) {
+		return nil, nil
+	} else if err == nil || os.IsExist(err) {
+		projutil.PrintDeprecationWarning("Package manifests are deprecated. " +
+			"Run `operator-sdk bundle create --generate-only` to create operator metadata")
+		b, err := ioutil.ReadFile(path)
+		if err != nil {
+			return nil, fmt.Errorf("failed to read package manifest %s: %v", path, err)
+		}
+		if err = yaml.Unmarshal(b, &pkg); err != nil {
+			return nil, fmt.Errorf("failed to unmarshal package manifest %s: %v", path, err)
+		}
+	} else {
+		return nil, fmt.Errorf("error reading package manifest %s: %v", path, err)
 	}
 
 	g.setChannels(&pkg)
 	sortChannelsByName(&pkg)
 
 	if err := validatePackageManifest(&pkg); err != nil {
-		log.Error(err)
-		os.Exit(1)
+		return nil, err
 	}
 
 	b, err := yaml.Marshal(pkg)
@@ -139,17 +156,18 @@ func (g pkgGenerator) generate() (map[string][]byte, error) {
 
 // buildPackageManifest will create a registry.PackageManifest from scratch, or reads
 // an existing one if found at the expected path.
+// Deprecated: only used for testing other methods on g.
 func (g pkgGenerator) buildPackageManifest() (registry.PackageManifest, error) {
+	pkgManifestOutputDir := filepath.Join(g.OutputDir, OLMCatalogChildDir, g.OperatorName)
+	path := filepath.Join(pkgManifestOutputDir, g.fileName)
 	pkg := registry.PackageManifest{}
-	olmCatalogDir := filepath.Join(g.OutputDir, OLMCatalogChildDir)
-	existingPkgManifest := filepath.Join(olmCatalogDir, g.OperatorName, g.fileName)
-	if isFileExist(existingPkgManifest) {
-		b, err := ioutil.ReadFile(existingPkgManifest)
+	if isFileExist(path) {
+		b, err := ioutil.ReadFile(path)
 		if err != nil {
-			return pkg, fmt.Errorf("failed to read package manifest %s: %v", existingPkgManifest, err)
+			return pkg, fmt.Errorf("failed to read package manifest %s: %v", path, err)
 		}
 		if err = yaml.Unmarshal(b, &pkg); err != nil {
-			return pkg, fmt.Errorf("failed to unmarshal package manifest %s: %v", existingPkgManifest, err)
+			return pkg, fmt.Errorf("failed to unmarshal package manifest %s: %v", path, err)
 		}
 	} else {
 		pkg = newPackageManifest(g.OperatorName, g.channel, g.csvVersion)
@@ -187,7 +205,8 @@ func validatePackageManifest(pkg *registry.PackageManifest) error {
 	return nil
 }
 
-// newPackageManifest will return the registry.PackageManifest populated
+// newPackageManifest will return the registry.PackageManifest populated.
+// Deprecated: only used for testing other methods on g.
 func newPackageManifest(operatorName, channelName, version string) registry.PackageManifest {
 	// Take the current CSV version to be the "alpha" channel, as an operator
 	// should only be designated anything more stable than "alpha" by a human.